uploaded-to-github
All images are 3D objects. Using 3D's max, I walked the lines along the contour and converted it to a polygon. Each polygon is colored according to the color of the selected sphere. For this mechanic, I used Ray and Raycast. when you hover the sphere over an object, it is colored with its color, and if you let go of the sphere, the polygon is colored and stores the value in accordance with the color of the used sphere. If you move it outside of the polygon, it returns the original color. Also implemented image magnification by double-clicking, for easy access to small details. Only 2 scenes were created. The first scene is the menu, the second is the game scene. When you select an image from the menu in the second scene, it is displayed according to the specified array element. The index of the array is written to the memory buffer for convenience, and is cleared when the scene is started.